requirements.txt
pocketbase
python-fasthtml
